Replace layer style example with a GeoZarr NDVI example
Computes an NDVI visualization for the red and NIR bands of a
Web-Optimized GeoZarr reflectance datacube via getLayerOptions.
The bands are identified through the eo:common_name band metadata
of the STAC asset instead of being hard-coded.
